requirements.txt
tqdm
sklearn
scipy
seaborn
networkx
keras
tensorflow-macos